style

+++ bright.h
@@ -87,7 +87,7 @@ static const struct Command commands[] = {
     [COMMAND_HELP]  = {"-h", "--help",  "print this help message"  },
 };
 
-#define ARRAY_LENGTH(x) (sizeof(x) / sizeof(x[0]))
+#define ARRAY_LENGTH(x) (int)(sizeof(x) / sizeof(x[0]))
 
 #define NLEVELS 11
 extern char *program;
